fix: remove some `any`  (#547)
## Description
There are many uses of `any` throughout the codebase. We could aim to
tighten up the typings to be more precise. This PR makes only small
improvements, but please share any thoughts on further changes we could
make or naming conventions to standardize on. I want you to know that
all feedback is welcome. <3

@Jackie1210 @shuding.